pyproject.toml
[project]
name = "blog-app"
version = "0.1.0"
description = "FastAPI 블로그 API (클린 아키텍처 예제)"
readme = "README.md"
authors = [
    { name = "kspark-sqisoft", email = "ks.park@sqisoft.com" }
]
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=5.0.0",
    "email-validator>=2.3.0",
    "fastapi>=0.135.3",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.13.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.24",
    "sqlalchemy>=2.0.49",
    "uvicorn[standard]>=0.44.0",
]

[project.scripts]
blog-serve = "blog_app.__main__:main"
blog-seed = "blog_app.seed:main"

[build-system]
requires = ["uv_build>=0.9.5,<0.10.0"]
build-backend = "uv_build"
